# minproj/static/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'MinProj'
  version: ''
paths:
...
  /departments/:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get a department's list by token
      operationId: Departments
      description: |
        # Список всех ведомств
        ## Любой пользователь
      parameters: [ ]
      responses:
        '200':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - departments
    post:
      security:
        - bearerAuth: [ ]
      operationId: CreateDepartment
      summary: Create a department by administrator
      description: |
        # Создание ведомства
        ## Администратор
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Министерство здравоохранения
            required: true
      responses:
        '201':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - departments
...
  /locations/:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get a location's list by token
      operationId: Locations
      description: |
        # Список всех учреждений
        ## Администратор - все учреждения
        ## Операторы (валидные и активные) - все учреждения, кроме неактивных учреждений других ведомств (с которыми оператор не связан)
        ## Координаторы (валидные и активные) - активные учреждения
        ## Получатели (валидные и активные) - активные учреждения
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search by name
        - in: query
          name: ordering
          schema:
            type: string
          description: Order by id, name, address, email, phone, is_active, department_id
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
        - in: query
          name: offset
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set
        - in: query
          name: page
          schema:
            type: integer
          description: The page of items
        - in: query
          name: is_active
          schema:
            type: boolean
          description: filter by is_active
      responses:
        '200':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - locations
    post:
      security:
        - bearerAuth: [ ]
      operationId: CreateLocation
      summary: Create a location
      description: |
        # Создание учреждения
        ## Администратор - создание любого учреждения
        ## Операторы (валидные и активные) - создание учреждения собственного ведомства
        ## В request body перечислены все поля (ненужные удалить)
        ### Обязательные поля: "name", "address", "phone", "department_id" (это поле только для Админа, для Оператора оно будет игнорироваться)
        ### Уникальные поля: "name", "phone"
        ### Проверка на валидацию: "phone"
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocation'
            required: true
      responses:
        '201':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - locations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateLocation:
      type: object
      properties:
        name:
          type: string
          example: test1
        address:
          type: string
          example: Penza
#        lat:
#          type: number
#          example: 53.178386
#        lon:
#          type: number
#          example: 45.008664
        photo:
          type: string
          example: http://116.202.240.109:8080/media/7c0ccda4-9b1f-480d-b025-143766289624.jpg
        phone:
          type: string
          example: "+79998887766"
        email:
          type: string
          example: test@test.test
        department_id_id:
          type: integer
          example: 1
        is_active:
          type: boolean
...